fix(seq): address PR review on arrange + fast/slow modifiers
Negative fast/slow factor: docs claimed "negatives reverse time" but the
combinator reverses the query span, which yields silence (there is no
reversal combinator). Correct the docs in FactorPayload, TimeModifiable,
and the generated tooltip to "non-positive factor yields silence";
regenerate schemas.json. Add tests pinning silence for const and
pattern-string negative factors.

Drop the redundant JSON-unreachable is_finite() guard in
SectionCycles::to_finite_cycles (the fract/<=0 checks already reject
NaN/+-inf); keep the load-bearing one in lower_factor (Fraction::from maps
non-finite to 0, so without it a non-finite factor is silently accepted)
and exercise it with a direct test.

Add arrange combinator coverage: single finite section (cycle budget is a
no-op), drop-sections-after-infinite-tail, and zero-sum silence.

Widen pattern_idx u8 -> u32 across the highlight-slot machinery (FlatSpan,
walk_inner, Rebased, offset_pattern_idx), removing the ~255-source
highlight ceiling. FlatSpan stays 12 bytes (matches start/end); the field
is internal cache, read only as a Vec index, so there is no wire/schema
change.

+++ crates/modular_core/src/pattern_system/hap.rs
@@ -274,7 +274,7 @@ pub enum ArenaHapContext<'b> {
     /// highlights land in the right slot of the flat `per_source` list.
     /// Nests additively: `Rebased(a, Rebased(b, inner))` emits at `a + b + …`.
     Rebased {
-        base: u8,
+        base: u32,
         inner: &'b ArenaHapContext<'b>,
     },
 }
@@ -339,10 +339,10 @@ impl<'b> ArenaHapContext<'b> {
     ///
     /// A `Stripped` node folds its modifier-side spans back into the source
     /// pattern_idx for the rest of the subtree.
-    pub fn walk<F: FnMut(u8, &SourceSpan)>(&self, emit: &mut F) {
+    pub fn walk<F: FnMut(u32, &SourceSpan)>(&self, emit: &mut F) {
         // Pattern index counter — incremented when we step into a modifier
         // subtree at the root level.
-        let mut next_pattern_idx: u8 = 0;
+        let mut next_pattern_idx: u32 = 0;
         self.walk_inner(0, false, &mut next_pattern_idx, emit);
     }
 
@@ -352,10 +352,10 @@ impl<'b> ArenaHapContext<'b> {
     // `walk_inner::<closure>`).
     fn walk_inner(
         &self,
-        pattern_idx: u8,
+        pattern_idx: u32,
         stripped: bool,
-        next_pattern_idx: &mut u8,
-        emit: &mut dyn FnMut(u8, &SourceSpan),
+        next_pattern_idx: &mut u32,
+        emit: &mut dyn FnMut(u32, &SourceSpan),
     ) {
         match self {
             ArenaHapContext::Empty => {}
@@ -387,7 +387,7 @@ impl<'b> ArenaHapContext<'b> {
                 // the outer counter is advanced past `base + width` so a
                 // sibling modifier chain continues from the right slot.
                 let base = *base;
-                let mut local_next: u8 = 0;
+                let mut local_next: u32 = 0;
                 inner.walk_inner(0, stripped, &mut local_next, &mut |idx, span| {
                     emit(base.saturating_add(idx), span);
                 });
